chore: reply smart pay review
